requirements.txt
anyio==4.4.0
asgiref==3.8.1
astor==0.8.1
attrdict==2.0.1
Babel==2.15.0
bce-python-sdk==0.9.17
beautifulsoup4==4.12.3
blend-modes==2.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
Django==4.2.13
djangorestframework==3.15.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.53.0
fsspec==2024.6.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
layoutparser==0.3.4
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.2.2
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
portalocker==2.10.0
premailer==3.10.0
protobuf==3.20.2
psutil==6.0.0
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
PyMuPDF==1.24.6
PyMuPDFb==1.24.6
pyparsing==3.1.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.9.3
rarfile==4.2
regex==2024.5.15
reportlab==4.2.2
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.13.0
termcolor==2.4.0
tifffile==2024.6.18
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visualdl==2.5.3
Werkzeug==3.0.3
zipp==3.19.2
